OrgImage Display
Modify orgImageDisplay so that it is more centralized and that navigating to orgImageScreen won't show default only
